Move fakeResponse to fixtures.py
